requirements.txt
langchain==0.3.25
langchain-core==0.3.63
langchain-google-genai==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
PyYAML==6.0.2
streamlit==1.45.1
python-dotenv==1.1.0